Makes data_source variable optional
Many projects don't have a data_source of s3, nfs or url type.
This allows the creation of a project without data_source.
In the future, it might be useful to add commented lines in the Makefile
with the omitted commands, in the case the user decides to use one of
the implemented data sources.